Unrelated:

1. Fixed tracy (again the cmake interface vs public issue)
2. Moved some APIs from `Vec` to `VecView`.

Optimizations:

1. Only do collision check once, use thread local buffers to store the result and merge in parallel when everything is completed. Note that this requires making the vector copy sequential, so I added some flags to help.
2. Allow reusing collider result buffer, reduce reallocation during complex polygon triangulation.
3. Some low level optimization.

This results in about 8% time reduction for `perfTest` (2.5s to 2.3s), and 10% time reduction for running triangulation tests (~1900ms to ~1700ms)