@pca006132 2024-10-09
Just got some time, I still don't understand why the previous test failure can happen, but I guess it is caused by the cmake changes in the master branch. The weird thing is that I haven't merged the changes at [f632177](https://github.com/elalish/manifold/pull/978/commits/f6321777acb4d0c082907256b8fb16696d616d5f). Anyway, the tests pass now.

For the black formatter error, I think this is caused by the latest version (released a day ago) no longer supports python 3.8, and Ubuntu 20.04 ship with python 3.8 by default. Should be fixed by installing python 3.12. (which I did previously for the cmake formatter because I forgot Ubuntu ships with python preinstalled...)